Add ruby management to base role
Installing bundler is a preparatory step for the numerous ruby web apps
my server is going to host.

Note that the nginx-passenger role does not depend on base. If ruby is
not installed, asking pacman to install the package will also bring in
ruby as a dependency. But it's better to explicitly install ruby anyway
so that it doesn't get torn out by a pacman -Rs uninstallation.