Enhance nginx templates
This commit introduces a really sweet new nginx configuration template.
It's just as elegant as Benno Joy's but actually more powerful. It also
implements all the associated changes needed to take advantage of the
new capabilities that this template provides.

We also have to move all the things that are referring to nginx_user,
since those are now under nginx_conf.user.