Add stringer role
Recently I tore down my entire installation of Firefox 33 and built it
from the ground up again. As a consequence, I seem to have lost my RSS
reader. Installing stringer became a significantly higher priority as a
result.

stringer's adherence to 12factor is a bit inconvenient for this system,
but we can handle it via passenger_set_cgi_param. The only real problem
is the fixed Ruby version (2.0.0) in the Gemfile. We can get around that
the same way stringer's own travis tests get around it - just strip the
top line off.